+++ src/termout.c
@@ -707,9 +707,11 @@ static uint
           sum += 0x80;
         if (attr & ATTR_INVISIBLE) {
           sum += 0x08;
-          // for xterm, invisible char value is apparently always 0x20
+#ifdef xterm_before_390
+          // fixed in xterm 390: invisible char value was always 0x20
           sum -= line->chars[x].chr;
           sum += ' ';
+#endif
         }
         if (attr & ATTR_PROTECTED)
           sum += 0x04;
@@ -4131,6 +4133,8 @@ static void
         child_printf("\eP1$r%u$}\e\\", term.st_active);
       } else if (!strcmp(s, "-p")) {  // DECARR (auto repeat rate)
         child_printf("\eP1$r%u-p\e\\", term.repeat_rate);
+      } else if (!strcmp(s, ">4m")) {  // XTQMODKEYS
+        child_printf("\eP1$r>4;%um\e\\", term.modify_other_keys);
       } else {
         child_printf("\eP0$r\e\\");
       }