+++ src/wintext.c
@@ -3414,6 +3414,24 @@ void
   }
 #endif
 
+#ifdef debug_non_blank_lines
+  void printline() {
+    bool dopri = false;
+    for (int i = 0; i < len; i++)
+      if (text[i] != ' ') {
+        dopri = true;
+        break;
+      }
+    if (dopri) {
+      printf("%d:%d %06X %06X (def %d) %d", ty, tx, fg, bg, default_bg, len);
+      for (int i = 0; i < len; i++)
+        printf(" %02X", text[i]);
+      printf("\n");
+    }
+  }
+  printline();
+#endif
+
  /* Begin text output */
   int yt = y + (ff->row_spacing / 2) - (lattr == LATTR_BOT ? cell_height : 0);
   int xt = x + (ff->col_spacing / 2);
@@ -3820,8 +3838,14 @@ draw:;
   if (combining || combining_double)
     *dxs = char_width;  // convince Windows to apply font underlining
 
+ /* Now, really draw the text */
+  // handle invisible and blinking attributes on image background
+  if (fg == bg && default_bg && *cfg.background)
+    goto skip_drawing;
+
   text_out_start(dc, text, len, dxs);
-  for (int xoff = 0; xoff < xwidth; xoff++)
+
+  for (int xoff = 0; xoff < xwidth; xoff++) {
     if (combining || combining_double) {
       // Workaround for mangled display of combining characters;
       // Arabic shaping should not be affected as the transformed 
@@ -3868,8 +3892,27 @@ draw:;
         overwropt = 0;
       }
     }
+
+    /*
+      With image background (-o Background=...png), if output in the 
+      top line begins with reverse or coloured background, 
+      a mysterious rendering bug hides the first chunk of output 
+      in frequent cases at that position.
+      (This was traced down in mintty deeply so the remaining suspicion 
+      is it's a bug in Windows.)
+      As a workaround, we invalidate the top-left cell right away 
+      so it gets printed to the window repeatedly, which effectively 
+      makes it visible from the first retry.
+     */
+    if (!tx && !ty && *cfg.background && !default_bg)
+      term_invalidate(0, 0, 0, 0);
+  }
+
   text_out_end();
 
+skip_drawing:;
+
+ /* Reset coordinate transformation */
   if (coord_transformed2) {
     SetWorldTransform(dc, &old_xform2);
     // restore these in case we're in a shadow loop